requirements.txt
torch==2.4.0+cu124
torch-geometric==2.6.1 #关于pyg的安装，可以参考这个指南：https://pytorch-geometric.readthedocs.io/en/2.5.2/notes/installation.html  
torch_cluster==1.6.3+pt24cu124
torch_scatter==2.1.2+pt24cu124
torch_sparse==0.6.18+pt24cu124
torch_spline_conv==1.2.2+pt24cu124
torchmetrics==1.6.0
torchvision==0.19.0+cu124
transformers==4.47.1
typing_extensions==4.12.2
tzdata==2024.2
ultralytics==8.3.32
ultralytics-thop==2.0.11
yarl==1.18.3
timm==0.9.7
tokenizers==0.21.0
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
monai==1.4.0
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
lightning==2.4.0
lightning-utilities==0.11.9
huggingface-hub==0.28.1

